completed methods created to correctly handle closing all accounts when deleting a customer without breaking or raising an error when there are no accounts, also all unit tests still passing with new tests added and also succeeding

+++ BankingApp/SAL/TransactionSAL/TransactionSALImplementation.py
@@ -76,6 +76,10 @@ class TransactionSALImplementation(TransactionSALInterface):
 
     def service_delete_all_transactions(self, account_id: int) -> bool:
         logging.info("Beginning SAL method delete all transactions")
-        result = self.transaction_dao.delete_all_transactions(account_id)
-        logging.info("Finishing SAL method delete all transactions")
-        return result
+        if type(account_id) != int:
+            logging.warning("SAL method delete all transactions, account ID not an integer")
+            raise FailedTransaction("The account ID field musts be an integer, please try again!")
+        else:
+            result = self.transaction_dao.delete_all_transactions(account_id)
+            logging.info("Finishing SAL method delete all transactions")
+            return result
